ibm1130 - Compiler cleanup

+++ Ibm1130/ibm1130_defs.h
@@ -27,7 +27,7 @@
 #define MAX(a,b)  (((a) >= (b)) ? (a) : (b))
 
 #ifndef _WIN32
-   int strnicmp (const char *a, const char *b, int n);
+   int strnicmp (const char *a, const char *b, size_t n);
    int strcmpi  (const char *a, const char *b);
 #endif
